// Yönetici Veri Paneli/src/app/model/bildirim.ts
export class Bildirim {
    key: string;
    baslik: string;
    icerik: string;
    uid: string;
    kayTarih: string;
    duzTarih: string;
    adsoyad: string;
    yetkinlik:string;
    mail: string;
    tel:string;
    adres:string;
    islem:boolean;
    kategori:string;
    bildirim:string;

}